backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.16.2
numpy==1.26.0
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
groq==1.1.2
python-dotenv==1.0.0
